+++ tests/unit/drivers/rerank/test_cohere_rerank_driver.py
@@ -1,0 +1,3 @@
+class TestCohereRerankDriver:
+    def test_run(self):
+        assert False